updated files

+++ CarND-Extended-Kalman-Filter-Project/src/kalman_filter.cpp
@@ -57,10 +57,10 @@ void KalmanFilter::Update(const VectorXd &z) {
 
 void KalmanFilter::UpdateEKF(const VectorXd &z) {
 
-  float px = x_(0);
-  float py = x_(1);
-  float vx = x_(2);
-  float vy = x_(3);
+  double px = x_(0);
+  double py = x_(1);
+  double vx = x_(2);
+  double vy = x_(3);
 
   float rho = sqrt(px*px+py*py);
   float phi = atan2(py,px);
